Type Alias rgb::alt::BGRA8

pub type BGRA8 = BGRA<u8>;

8-bit BGRA

Aliased Type

struct BGRA8 {
    pub b: u8,
    pub g: u8,
    pub r: u8,
    pub a: u8,
}

Fields
b: u8

Blue first

g: u8

Green

r: u8

Red

a: u8

Alpha last

Implementations

impl<T> BGRA<T>

pub const fn new(r: T, g: T, b: T, a: T) -> Self

๐Deprecated: This function has a misleading order of arguments. Use BGRA{} literal instead

Convenience function for creating a new pixel Warning: The order of arguments is R,G,B,A

impl<T, A> BGRA<T, A>

pub const fn new_alpha(r: T, g: T, b: T, a: A) -> Self

๐Deprecated: This function has a misleading order of arguments. Use BGRA{} literal instead

Convenience function for creating a new pixel Warning: The order of arguments is R,G,B,A

impl<T, A> BGRA<T, A>

pub fn rgb_mut(&mut self) -> &mut BGR<T>

๐Deprecated: This function will change. Use bgr_mut()

Provide a mutable view of only RGB components (leaving out alpha). Useful to change color without changing opacity.

pub fn bgr_mut(&mut self) -> &mut BGR<T>

Provide a mutable view of only RGB components (leaving out alpha). Useful to change color without changing opacity.

impl<T: Clone, A> BGRA<T, A>

pub fn rgb(&self) -> BGR<T>

๐Deprecated: This function will change. Use bgr()

Copy RGB components out of the RGBA struct

Note: you can use .into() to convert between other types

impl<T: Clone> BGRA<T>

pub fn iter(&self) -> Cloned<Iter<'_, T>>

Iterate over all components (length=4)

impl<T: Clone, A> BGRA<T, A>

pub fn bgr(&self) -> BGR<T>

Copy RGB components out of the RGBA struct

Note: you can use .into() to convert between other types

impl<T: Copy, A: Clone> BGRA<T, A>

pub fn map_rgb<F, U, B>(&self, f: F) -> BGRA<U, B>where F: FnMut(T) -> U, U: Clone, B: From<A> + Clone,

Create new RGBA with the same alpha value, but different RGB values

pub fn alpha(&self, a: A) -> Self

Create a new RGBA with the new alpha value, but same RGB values

pub fn map_alpha<F, B>(&self, f: F) -> BGRA<T, B>where F: FnOnce(A) -> B,

Create a new RGBA with a new alpha value created by the callback. Allows changing of the type used for the alpha channel.