rgb::alt

Type Alias GRAYA16

pub type GRAYA16 = GrayAlpha<u16>;

👎Deprecated: Refer to ::rgb::alt::GrayAlpha<u16> directly (this type alias will change in the next major version)

16-bit gray with alpha in machine’s native endian

Aliased Type

#[repr(C)]
pub struct GRAYA16(pub u16, pub u16);

Tuple Fields
0: u16
👎Deprecated: Please use the .v field instaed (it’s available through the magic of Deref to GrayA type)

Grayscale Component

This field has been renamed to .v

1: u16
👎Deprecated: Please use the .a field instead (it’s available through the magic of Deref to GrayA type)

Alpha Component. This field has been renamed to .a.