rgb

Struct Rgb

#[repr(C)]
pub struct Rgb<T> {
    pub r: T,
    pub g: T,
    pub b: T,
}

A Red + Green + Blue pixel.

Examples

use rgb::Rgb;

let pixel: Rgb<u8> = Rgb { r: 0, g: 0, b: 0 };

Fields
r: T

Red Component

g: T

Green Component

b: T

Blue Component

Implementations

impl<T: Copy> Rgb<T>

pub const fn new(red: T, green: T, blue: T) -> Self

Creates a new Rgb pixel type from its components.

Alternatively, you can use struct literal syntax to create the new pixel type:

use rgb::Rgb;

let pixel = Rgb {r : red, g : green, b : blue};

impl<T: Clone> Rgb<T>

pub fn iter(&self) -> Cloned<Iter<'_, T>>

Iterate over color components (R, G, and B)

impl<T: Clone> Rgb<T>

pub fn with_alpha(&self, a: T) -> RGBA<T>

Convenience function for converting to RGBA

pub fn new_alpha<A>(&self, a: A) -> RGBA<T, A>

๐Deprecated: use .with_alpha(a) instead

Convenience function for converting to RGBA with alpha channel of a different type than type of the pixels