Type Alias rgb::RGB8

pub type RGB8 = RGB<u8>;

8-bit RGB

The colorspace is technically undefined, but generally sRGB is assumed.

Aliased Type

struct RGB8 {
    pub r: u8,
    pub g: u8,
    pub b: u8,
}

Fields
r: u8

Red

g: u8

Green

b: u8

Blue

Implementations

impl<T> RGB<T>

pub const fn new(r: T, g: T, b: T) -> Self

Convenience function for creating a new pixel The order of arguments is R,G,B

impl<T: Clone> RGB<T>

pub fn iter(&self) -> Cloned<Iter<'_, T>>

Iterate over color components (R, G, and B)

impl<T: Clone> RGB<T>

pub fn alpha(&self, a: T) -> RGBA<T>

Convenience function for converting to RGBA

pub fn new_alpha<A>(&self, a: A) -> RGBA<T, A>

Convenience function for converting to RGBA with alpha channel of a different type than type of the pixels