Type Alias rgb::RGBA16

pub type RGBA16 = RGBA<u16>;

16-bit RGB in machine’s native endian. 0 = transparent, 65535 = opaque.

Alpha is last.

Aliased Type

struct RGBA16 {
    pub r: u16,
    pub g: u16,
    pub b: u16,
    pub a: u16,
}

Fields
r: u16

Red

g: u16

Green

b: u16

Blue

a: u16

Alpha

Implementations

impl<T> RGBA<T>

pub const fn new(r: T, g: T, b: T, a: T) -> Self

Convenience function for creating a new pixel The order of arguments is R,G,B,A

impl<T, A> RGBA<T, A>

pub const fn new_alpha(r: T, g: T, b: T, a: A) -> Self

Convenience function for creating a new pixel The order of arguments is R,G,B,A

impl<T, A> RGBA<T, A>

pub fn rgb_mut(&mut self) -> &mut RGB<T>

Provide a mutable view of only RGB components (leaving out alpha). Useful to change color without changing opacity.

impl<T: Clone, A> RGBA<T, A>

pub fn rgb(&self) -> RGB<T>

Copy RGB components out of the RGBA struct

Note: you can use .into() to convert between other types

impl<T: Clone> RGBA<T>

pub fn iter(&self) -> Cloned<Iter<'_, T>>

Iterate over all components (length=4)

impl<T: Clone, A> RGBA<T, A>

pub fn bgr(&self) -> BGR<T>

Copy RGB components out of the RGBA struct

Note: you can use .into() to convert between other types

impl<T: Copy, A: Clone> RGBA<T, A>

pub fn map_rgb<F, U, B>(&self, f: F) -> RGBA<U, B>where F: FnMut(T) -> U, U: Clone, B: From<A> + Clone,

Create new RGBA with the same alpha value, but different RGB values

pub fn alpha(&self, a: A) -> Self

Create a new RGBA with the new alpha value, but same RGB values

pub fn map_alpha<F, B>(&self, f: F) -> RGBA<T, B>where F: FnOnce(A) -> B,

Create a new RGBA with a new alpha value created by the callback. Allows changing of the type used for the alpha channel.