Struct rgb::RGBA

#[repr(C)]

pub struct RGBA<ComponentType, AlphaComponentType = ComponentType> {
    pub r: ComponentType,
    pub g: ComponentType,
    pub b: ComponentType,
    pub a: AlphaComponentType,
}

This is it. The component type can be u8 (aliased as RGBA8), u16 (aliased as RGBA16), or any other type (but simple copyable types are recommended.)

You can specify a different type for alpha, but it's only for special cases (e.g. if you use a newtype like RGBA, u16>).

Fields
r: ComponentType

Red

g: ComponentType

Green

b: ComponentType

Blue

a: AlphaComponentType

Alpha

Methods

impl<T: Clone> RGBA<T>

fn new(r: T, g: T, b: T, a: T) -> Self

fn iter(&self) -> Cloned<Iter<T>>

Iterate over all components (length=4)

impl<T: Clone, A> RGBA<T, A>

fn rgb(&self) -> RGB<T>

Copy RGB components out of the RGBA struct

impl<T, A> RGBA<T, A>

fn rgb_mut(&mut self) -> &mut RGB<T>

Provide a mutable view of only RGB components (leaving out alpha). Useful to change color without changing opacity.

impl<T: Copy, A: Clone> RGBA<T, A>

fn map_rgb<F, U, B>(&self, f: F) -> RGBA<U, B> where
F: FnMut(T) -> U,
U: Clone,
B: From<A> + Clone,

Create new RGBA with the same alpha value, but different RGB values