Struct rgb::RGBA

#[repr(C)]
pub struct RGBA<T, A = T> {
    pub r: T,
    pub g: T,
    pub b: T,
    pub a: A,
}

An Red + Green + Blue + Alpha pixel.

Fields
r: T

Red Component

g: T

Green Component

b: T

Blue Component

a: A

Alpha Component

Implementations

impl<T> RGBA<T>

pub const fn new(r: T, g: T, b: T, a: T) -> Self

Convenience function for creating a new pixel The order of arguments is R,G,B,A

impl<T, A> RGBA<T, A>

pub const fn new_alpha(r: T, g: T, b: T, a: A) -> Self

Convenience function for creating a new pixel The order of arguments is R,G,B,A

impl<T, A> RGBA<T, A>

pub fn rgb_mut(&mut self) -> &mut RGB<T>

Provide a mutable view of only RGB components (leaving out alpha). Useful to change color without changing opacity.

impl<T: Clone, A> RGBA<T, A>

pub fn rgb(&self) -> RGB<T>

Copy RGB components out of the RGBA struct

Note: you can use .into() to convert between other types

impl<T: Clone> RGBA<T>

pub fn iter(&self) -> Cloned<Iter<'_, T>>

Iterate over all components (length=4)

impl<T: Clone, A> RGBA<T, A>

pub fn bgr(&self) -> BGR<T>

Copy RGB components out of the RGBA struct

Note: you can use .into() to convert between other types

impl<T: Copy, A: Clone> RGBA<T, A>

pub fn map_rgb<F, U, B>(&self, f: F) -> RGBA<U, B>
where F: FnMut(T) -> U, U: Clone, B: From<A> + Clone,

Create new RGBA with the same alpha value, but different RGB values

pub fn alpha(&self, a: A) -> Self

Create a new RGBA with the new alpha value, but same RGB values

pub fn map_alpha<F, B>(&self, f: F) -> RGBA<T, B>
where F: FnOnce(A) -> B,

Create a new RGBA with a new alpha value created by the callback. Allows changing of the type used for the alpha channel.