Struct rgb::RGBA

#[repr(C)]
pub struct RGBA<ComponentType, AlphaComponentType = ComponentType> {
    pub r: ComponentType,
    pub g: ComponentType,
    pub b: ComponentType,
    pub a: AlphaComponentType,
}

The RGBA pixel

The component type can be u8 (aliased as RGBA8), u16 (aliased as RGBA16), or any other type (but simple copyable types are recommended.)

You can specify a different type for alpha, but it's only for special cases (e.g. if you use a newtype like RGBA<LinearLight<u16>, u16>).

Fields
r: ComponentType

Red

g: ComponentType

Green

b: ComponentType

Blue

a: AlphaComponentType

Alpha

Methods

impl<T> RGBA<T>

pub const fn new(r: T, g: T, b: T, a: T) -> Self

Convenience function for creating a new pixel The order of arguments is R,G,B,A

impl<T, A> RGBA<T, A>

pub const fn new_alpha(r: T, g: T, b: T, a: A) -> Self

Convenience function for creating a new pixel The order of arguments is R,G,B,A

impl<T: Clone> RGBA<T>

pub fn iter(&self) -> Cloned<Iter<T>>

Iterate over all components (length=4)

impl<T: Clone, A> RGBA<T, A>

pub fn rgb(&self) -> RGB<T>

Copy RGB components out of the RGBA struct

Note: you can use .into() to convert between other types

impl<T, A> RGBA<T, A>

pub fn rgb_mut(&mut self) -> &mut RGB<T>

Provide a mutable view of only RGB components (leaving out alpha). Useful to change color without changing opacity.

impl<T: Copy, A: Clone> RGBA<T, A>

pub fn map_rgb<F, U, B>(&self, f: F) -> RGBA<U, B> where
F: FnMut(T) -> U,
U: Clone,
B: From<A> + Clone,

Create new RGBA with the same alpha value, but different RGB values

pub fn alpha(&self, a: A) -> Self

Create a new RGBA with the new alpha value, but same RGB values

pub fn map_alpha<F, B>(&self, f: F) -> RGBA<T, B> where
F: FnOnce(A) -> B,

Create a new RGBA with a new alpha value created by the callback. Allows changing of the type used for the alpha channel.