rgb

Trait ColorComponentMap

pub trait ColorComponentMap<DestPixel, SrcComponent, DestComponent> {
    // Provided methods
    fn map_colors<Callback>(&self, f: Callback) -> DestPixel
       where Callback: FnMut(SrcComponent) -> DestComponent { ... }
    fn map_c<Callback>(&self, f: Callback) -> DestPixel
       where Callback: FnMut(SrcComponent) -> DestComponent { ... }
}

Same as ComponentMap, but doesn’t change the alpha channel (if there’s any alpha).

Import via use rgb::prelude::*; instead of directly.

Provided Methods

fn map_colors<Callback>(&self, f: Callback) -> DestPixel
where Callback: FnMut(SrcComponent) -> DestComponent,

Convenience function for applying the same formula to every rgb/gray component, but skipping the alpha component.

Note that it returns the pixel directly, not an Interator.

fn map_c<Callback>(&self, f: Callback) -> DestPixel
where Callback: FnMut(SrcComponent) -> DestComponent,

👎Deprecated: renamed to map_colors

Alias of map_colors

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.