Trait rgb::ComponentSlice

pub trait ComponentSlice<T> {
    fn as_slice(&self) -> &[T];

    fn as_mut_slice(&mut self) -> &mut [T];
}

Casting the struct to slices of its components

Required methods

fn as_slice(&self) -> &[T]

The components interpreted as an array, e.g. one RGB expands to 3 elements.

It's implemented for individual pixels as well as slices of pixels.

fn as_mut_slice(&mut self) -> &mut [T]

The components interpreted as a mutable array, e.g. one RGB expands to 3 elements.

It's implemented for individual pixels as well as slices of pixels.

Loading content...